kubevirt: move types to a standalone dir
