+++ frontend/packages/kubevirt-plugin/src/types/vm/index.ts
